requirements.txt
paho-mqtt
protobuf==3.20.3
